Largely completed receive transport abstractions.
Last up on the list is the transmit semantics. My current plan for
transport is to feature-gate different transmit functions behind `std`,
but that doesn't fit nicely with my current model. More thinking when I
have more sleep :)
