+++ uavcan/src/transport/mod.rs
@@ -8,6 +8,8 @@
 mod can;
 
 use crate::internal::InternalRxFrame;
+use crate::session::SessionManager;
+use crate::RxError;
 
 pub trait SessionMetadata {
     /// Create a fresh instance of session metadata;
@@ -21,4 +23,15 @@ pub trait SessionMetadata {
     /// Final check to see if transfer was successful.
     // TODO maybe this doesn't need a frame?
     fn is_valid(&self, frame: &InternalRxFrame) -> bool;
+}
+
+// TODO does this go with Node or stay here?
+/// Trait to be implemented for Node, declaring a transport implementation.
+pub trait Transport {
+    type Frame;
+
+    // TODO not sure if I can use lifetimes in my impls properly
+    /// Process a frame, returning the internal transport-independant representation,
+    /// or errors if invalid.
+    fn rx_process_frame<'a, S: SessionManager>(node: &mut crate::node::Node<S, Self>, frame: Self::Frame) -> Result<Option<InternalRxFrame<'a>>, RxError>;
 }
